Fixing the first round of PR review by Glen Whitney
